Move Dto's to Dto in include directory and created Dto namespace
